TT-1962: Implement upload certificates page
Quite a bit of duplication between the service and matching service
pages now, starting to be a bit of a nuisance.

Also it's pretty hacky the way this works - pages always show signing
certificates, even when the user uploads encryption certs. It's only a
prototype at the end of the day...